+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove long-deprecated functions
 - Correct `normalize_rect` (oops!)
 - Add and document `finalize` method on ContentObjects
+- Make `PageList` work more or less like a `Sequence`
 - TODO: Maximize test coverage
 - TODO: Optimize marked content section access
 - TODO: Add method to complete parent tree for page